+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/list_view/ChatMessageListCellFactory.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chat_messages.list_view;
+package bisq.desktop.main.content.chat.list_view;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
@@ -23,10 +23,10 @@ import bisq.desktop.main.content.bisq_easy.offerbook.MyOfferMessageBox;
 import bisq.desktop.main.content.bisq_easy.offerbook.PeerOfferMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.MyProtocolLogMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.PeerProtocolLogMessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.message_box.LeaveChatMessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.message_box.MessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.message_box.MyMessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.message_box.PeerMessageBox;
+import bisq.desktop.main.content.chat.list_view.message_box.LeaveChatMessageBox;
+import bisq.desktop.main.content.chat.list_view.message_box.MessageBox;
+import bisq.desktop.main.content.chat.list_view.message_box.MyMessageBox;
+import bisq.desktop.main.content.chat.list_view.message_box.PeerMessageBox;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
