+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/list_view/ChatMessagesListModel.java
@@ -1,4 +1,4 @@
-package bisq.desktop.main.content.components.chat_messages.list_view;
+package bisq.desktop.main.content.chat.list_view;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatChannelDomain;
