+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/list_view/message_box/MyMessageBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chat_messages.list_view.message_box;
+package bisq.desktop.main.content.chat.list_view.message_box;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
@@ -24,9 +24,9 @@ import bisq.desktop.common.Icons;
 import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.controls.BisqTextArea;
 import bisq.desktop.components.controls.BisqTooltip;
-import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessageListItem;
-import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessagesListController;
-import bisq.desktop.main.content.components.chat_messages.list_view.ChatMessagesListModel;
+import bisq.desktop.main.content.chat.list_view.ChatMessageListItem;
+import bisq.desktop.main.content.chat.list_view.ChatMessagesListController;
+import bisq.desktop.main.content.chat.list_view.ChatMessagesListModel;
 import bisq.i18n.Res;
 import bisq.network.p2p.services.confidential.ack.MessageDeliveryStatus;
 import de.jensd.fx.fontawesome.AwesomeDude;
@@ -122,7 +122,7 @@ public final class MyMessageBox extends BubbleMessageBox {
         deliveryState.getTooltip().textProperty().bind(item.getMessageDeliveryStatusTooltip());
         editInputField.maxWidthProperty().bind(message.widthProperty());
 
-        VBox.setMargin(deliveryStateHBox, new Insets(4, 0, -3, 0));
+        setMargin(deliveryStateHBox, new Insets(4, 0, -3, 0));
         messageHBox.getChildren().setAll(Spacer.fillHBox(), messageBgHBox);
 
         contentVBox.getChildren().setAll(userNameAndDateHBox, messageHBox, editButtonsHBox, deliveryStateHBox);
@@ -134,7 +134,7 @@ public final class MyMessageBox extends BubbleMessageBox {
 
         userNameAndDateHBox = new HBox(10, dateTime, userName);
         userNameAndDateHBox.setAlignment(Pos.CENTER_RIGHT);
-        VBox.setMargin(userNameAndDateHBox, new Insets(-5, 10, -5, 0));
+        setMargin(userNameAndDateHBox, new Insets(-5, 10, -5, 0));
     }
 
     @Override
@@ -161,7 +161,7 @@ public final class MyMessageBox extends BubbleMessageBox {
         editButtonsHBox = new HBox(15, Spacer.fillHBox(), cancelEditButton, saveEditButton);
         editButtonsHBox.setVisible(false);
         editButtonsHBox.setManaged(false);
-        VBox.setMargin(editButtonsHBox, new Insets(10, 25, -15, 0));
+        setMargin(editButtonsHBox, new Insets(10, 25, -15, 0));
         handleEditBox();
     }
